requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
geopandas==0.14.3
idna==3.6
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.1.0
